PR:478 Review changes. Combined QuestionnaireItemExtension files into single one. Changed localizedText and localizedPrefix to extension property instead of function. Updated tests appropriately.
