+++ datacapture/src/main/java/com/google/android/fhir/datacapture/MoreQuestionnaireItemComponents.kt
@@ -16,8 +16,11 @@
 
 package com.google.android.fhir.datacapture
 
+import java.util.*
 import org.hl7.fhir.r4.model.CodeableConcept
 import org.hl7.fhir.r4.model.Questionnaire
+import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.StringType
 
 internal const val ITEM_CONTROL_DROP_DOWN = "drop-down"
 internal const val ITEM_CONTROL_RADIO_BUTTON = "radio-button"
@@ -35,3 +38,27 @@ internal val Questionnaire.QuestionnaireItemComponent.itemControl: String?
       codeableConcept?.coding?.firstOrNull { it.system == EXTENSION_ITEM_CONTROL_SYSTEM }?.code
     return listOf(ITEM_CONTROL_DROP_DOWN, ITEM_CONTROL_RADIO_BUTTON).firstOrNull { it == code }
   }
+
+/**
+ * Whether the corresponding [QuestionnaireResponse.QuestionnaireResponseItemComponent] should have
+ * nested items within [QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent](s).
+ */
+val Questionnaire.QuestionnaireItemComponent.hasNestedItemsWithinAnswers: Boolean
+  get() = item.isNotEmpty() && type != Questionnaire.QuestionnaireItemType.GROUP
+
+private fun StringType.getLocalizedText(lang: String = Locale.getDefault().toLanguageTag()) : String?{
+  return getTranslation(lang)
+    ?: getTranslation(lang.split("-").first()) ?: value
+}
+
+/**
+ * Localized value of [Questionnaire.QuestionnaireItemComponent.text] if translation is present. Default value otherwise.
+ */
+internal val Questionnaire.QuestionnaireItemComponent.localizedText : String?
+  get() = textElement?.getLocalizedText()
+
+/**
+ * Localized value of [Questionnaire.QuestionnaireItemComponent.prefix] if translation is present. Default value otherwise.
+ */
+internal val Questionnaire.QuestionnaireItemComponent.localizedPrefix : String?
+  get() = prefixElement?.getLocalizedText()
