+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/QuestionnaireItemDropDownViewHolderFactory.kt
@@ -49,11 +49,11 @@ internal object QuestionnaireItemDropDownViewHolderFactory :
         this.questionnaireItemViewItem = questionnaireItemViewItem
         if (!questionnaireItemViewItem.questionnaireItem.prefix.isNullOrEmpty()) {
           prefixTextView.visibility = View.VISIBLE
-          prefixTextView.text = questionnaireItemViewItem.questionnaireItem.localizedPrefix()
+          prefixTextView.text = questionnaireItemViewItem.questionnaireItem.localizedPrefix
         } else {
           prefixTextView.visibility = View.GONE
         }
-        textView.text = questionnaireItemViewItem.questionnaireItem.localizedText()
+        textView.text = questionnaireItemViewItem.questionnaireItem.localizedText
         val answerOptionString =
           this.questionnaireItemViewItem.questionnaireItem.answerOption.map { it.displayString }
         val adapter =
