--- datacapture/src/test/java/com/google/android/fhir/datacapture/MoreQuestionnaireItemExtensionsTest.kt
@@ -1,120 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.android.fhir.datacapture
-
-import android.os.Build
-import com.google.common.truth.Truth.assertThat
-import org.hl7.fhir.r4.model.CodeableConcept
-import org.hl7.fhir.r4.model.Coding
-import org.hl7.fhir.r4.model.Extension
-import org.hl7.fhir.r4.model.Questionnaire
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-import org.robolectric.annotation.Config
-
-@RunWith(RobolectricTestRunner::class)
-@Config(sdk = [Build.VERSION_CODES.P])
-class MoreQuestionnaireItemExtensionsTest {
-
-  @Test
-  fun itemControl_shouldReturnItemControlCodeDropDown() {
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().setType(Questionnaire.QuestionnaireItemType.CHOICE)
-    questionnaireItem.addExtension(
-      Extension()
-        .setUrl(EXTENSION_ITEM_CONTROL_URL)
-        .setValue(
-          CodeableConcept()
-            .addCoding(
-              Coding()
-                .setCode(ITEM_CONTROL_DROP_DOWN)
-                .setDisplay("Drop Down")
-                .setSystem(EXTENSION_ITEM_CONTROL_SYSTEM)
-            )
-        )
-    )
-
-    assertThat(questionnaireItem.itemControl).isEqualTo(ITEM_CONTROL_DROP_DOWN)
-  }
-
-  @Test
-  fun itemControl_shouldReturnItemControlCodeRadioButton() {
-
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().setType(Questionnaire.QuestionnaireItemType.CHOICE)
-    questionnaireItem.addExtension(
-      Extension()
-        .setUrl(EXTENSION_ITEM_CONTROL_URL)
-        .setValue(
-          CodeableConcept()
-            .addCoding(
-              Coding()
-                .setCode(ITEM_CONTROL_RADIO_BUTTON)
-                .setDisplay("Radio Group")
-                .setSystem(EXTENSION_ITEM_CONTROL_SYSTEM)
-            )
-        )
-    )
-
-    assertThat(questionnaireItem.itemControl).isEqualTo(ITEM_CONTROL_RADIO_BUTTON)
-  }
-
-  @Test
-  fun itemControl_wrongExtensionUrl_shouldReturnNull() {
-
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().setType(Questionnaire.QuestionnaireItemType.CHOICE)
-    questionnaireItem.addExtension(
-      Extension()
-        .setUrl("null-test")
-        .setValue(
-          CodeableConcept()
-            .addCoding(
-              Coding()
-                .setCode(ITEM_CONTROL_DROP_DOWN)
-                .setDisplay("Drop Down")
-                .setSystem(EXTENSION_ITEM_CONTROL_SYSTEM)
-            )
-        )
-    )
-
-    assertThat(questionnaireItem.itemControl).isNull()
-  }
-
-  @Test
-  fun itemControl_wrongExtensionCoding_shouldReturnNull() {
-
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().setType(Questionnaire.QuestionnaireItemType.CHOICE)
-    questionnaireItem.addExtension(
-      Extension()
-        .setUrl(EXTENSION_ITEM_CONTROL_URL)
-        .setValue(
-          CodeableConcept()
-            .addCoding(
-              Coding()
-                .setCode("null-test")
-                .setDisplay("Drop Down")
-                .setSystem(EXTENSION_ITEM_CONTROL_SYSTEM)
-            )
-        )
-    )
-
-    assertThat(questionnaireItem.itemControl).isNull()
-  }
-}
