--- datacapture/src/test/java/com/google/android/fhir/datacapture/QuestionnaireItemTranslationExtensionTest.kt
@@ -1,134 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.android.fhir.datacapture
-
-import android.os.Build
-import com.google.common.truth.Truth.assertThat
-import java.util.Locale
-import org.hl7.fhir.r4.model.Extension
-import org.hl7.fhir.r4.model.Questionnaire
-import org.hl7.fhir.r4.model.StringType
-import org.hl7.fhir.r4.utils.ToolingExtensions
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-import org.robolectric.annotation.Config
-
-@RunWith(RobolectricTestRunner::class)
-@Config(sdk = [Build.VERSION_CODES.P])
-class QuestionnaireItemTranslationExtensionTest {
-  @Test
-  fun localizedValues_default() {
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().apply {
-        text = "Patient Information"
-        prefix = "One"
-      }
-
-    assertThat(questionnaireItem.localizedText()).isEqualTo("Patient Information")
-    assertThat(questionnaireItem.localizedPrefix()).isEqualTo("One")
-  }
-
-  @Test
-  fun localizedValues_vietnameseTranslationAndDefaultLocale_shouldGetDefault() {
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().apply {
-        text = "Patient Information"
-        textElement.apply {
-          addExtension(
-            Extension(ToolingExtensions.EXT_TRANSLATION).apply {
-              addExtension(Extension("lang", StringType("vi-VN")))
-              addExtension(Extension("content", StringType("Thông tin bệnh nhân")))
-            }
-          )
-        }
-
-        prefix = "One"
-        prefixElement.apply {
-          addExtension(
-            Extension(ToolingExtensions.EXT_TRANSLATION).apply {
-              addExtension(Extension("lang", StringType("vi-VN")))
-              addExtension(Extension("content", StringType("Một")))
-            }
-          )
-        }
-      }
-
-    assertThat(questionnaireItem.localizedText()).isEqualTo("Patient Information")
-    assertThat(questionnaireItem.localizedPrefix()).isEqualTo("One")
-  }
-
-  @Test
-  fun localizedValues_vietnameseTranslationAndVietnameseLocale_shouldGetTranslation() {
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().apply {
-        text = "Patient Information"
-        textElement.apply {
-          addExtension(
-            Extension(ToolingExtensions.EXT_TRANSLATION).apply {
-              addExtension(Extension("lang", StringType("vi-VN")))
-              addExtension(Extension("content", StringType("Thông tin bệnh nhân")))
-            }
-          )
-        }
-
-        prefix = "One"
-        prefixElement.apply {
-          addExtension(
-            Extension(ToolingExtensions.EXT_TRANSLATION).apply {
-              addExtension(Extension("lang", StringType("vi-VN")))
-              addExtension(Extension("content", StringType("Một")))
-            }
-          )
-        }
-      }
-
-    Locale.setDefault(Locale.forLanguageTag("vi-VN"))
-    assertThat(questionnaireItem.localizedText()).isEqualTo("Thông tin bệnh nhân")
-    assertThat(questionnaireItem.localizedPrefix()).isEqualTo("Một")
-  }
-
-  @Test
-  fun localizedValues_vietnameseTranslationAndVietnameseLocaleWithCountry_shouldGetTranslation() {
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().apply {
-        text = "Patient Information"
-        textElement.apply {
-          addExtension(
-            Extension(ToolingExtensions.EXT_TRANSLATION).apply {
-              addExtension(Extension("lang", StringType("vi")))
-              addExtension(Extension("content", StringType("Thông tin bệnh nhân")))
-            }
-          )
-        }
-
-        prefix = "One"
-        prefixElement.apply {
-          addExtension(
-            Extension(ToolingExtensions.EXT_TRANSLATION).apply {
-              addExtension(Extension("lang", StringType("vi")))
-              addExtension(Extension("content", StringType("Một")))
-            }
-          )
-        }
-      }
-
-    Locale.setDefault(Locale.forLanguageTag("vi-VN"))
-    assertThat(questionnaireItem.localizedText()).isEqualTo("Thông tin bệnh nhân")
-    assertThat(questionnaireItem.localizedPrefix()).isEqualTo("Một")
-  }
-}
